+++ debugging/timeout_spec.js
@@ -1,5 +1,4 @@
 describe('timeout possibilities', function() {
-  ptor = protractor.getInstance();
   jasmine.getEnv().defaultTimeoutInterval = 33;
 
 
@@ -8,16 +7,16 @@ describe('timeout possibilities', function() {
   });
 
   it('should timeout due to webdriver script timeout', function() {
-    ptor.driver.manage().timeouts().setScriptTimeout(55);
+    browser.driver.manage().timeouts().setScriptTimeout(55);
 
-    ptor.get('app/index.html#/form');
+    browser.get('app/index.html#/form');
 
-    ptor.driver.executeAsyncScript(function() {
+    browser.driver.executeAsyncScript(function() {
       var callback = arguments[arguments.length - 1];
       setTimeout(callback, 500);
     });
 
-    expect(ptor.findElement(protractor.By.binding('greeting')).getText()).
+    expect(element(by.binding('greeting')).getText()).
         toEqual('Hiya');
   }, 5000); // The 5000 here sets the Jasmine spec timeout.
 
@@ -29,8 +28,27 @@ describe('timeout possibilities', function() {
     expect(true).toEqual(true);
   });
 
+  describe('waitForAngular', function() {
+    it('should timeout and give a reasonable message', function() {
+
+      browser.driver.manage().timeouts().setScriptTimeout(55);
+
+      browser.get('app/index.html#/async');
+
+
+      var status = element(by.binding('slowHttpStatus'));
+      var button = element(by.css('[ng-click="slowHttp()"]'));
+
+      expect(status.getText()).toEqual('not started');
+
+      button.click();
+
+      expect(status.getText()).toEqual('done');
+    }, 5000); // The 5000 here sets the Jasmine spec timeout.
+  });
+
   it('should timeout due to Jasmine spec timeout', function() {
-    ptor.driver.sleep(1000);
+    browser.driver.sleep(1000);
     expect(true).toBe(true);
   });