+++ spec/onPrepare/startup.js
@@ -1,0 +1,1 @@
+browser.params.password = '12345';